Rename `sparkNodes.js` to `useSparkNodes.js`
